requirements.txt
Django==4.2.18
django-filter==24.3
djangorestframework==3.15.2
python-dotenv==1.0.1
